* trunk: (22 commits)
  exclude sboms from source distro
  try ant for jdk 20 build due to policeman issues
  try ant for jdk 20 build due to policeman issues
  add sboms to bin/src dits
  [github-408] Increase performance of StylesTable.putStyle. Thanks to Jason Mirra. This closes #408
  java 20 build
  try to build sboms
  archive xmlbeans jdk 1.19 build
  archive xmlbeans sboms
  cyclonedx per module
  upgrade cyclonedx
  revert accidental change to multiply/divide
  upgrade mockito and byte-buddy
  slf4j 2.0.6
  [bug-66319] do not create HSSF patriarch to find comments (only use one that pre-exists)
  untidy hack for shared string counts that are higher than Integer.MAX_VALUE
  untidy hack for shared string counts that are higher than Integer.MAX_VALUE
  remove unnecessary short casts for getCellStyleAt
  try to speed up SXSSFCell getColumnIndex
  small optimisation on getCellStyle
  ...

# Conflicts:
#	poi-ooxml/src/test/java/org/apache/poi/xslf/usermodel/TestXSLFTableCell.java

+++ poi-examples/src/main/java/org/apache/poi/examples/ss/html/ToHtml.java
@@ -320,7 +320,7 @@ public final class ToHtml {
 
     private String styleName(CellStyle style) {
         if (style == null) {
-            style = wb.getCellStyleAt((short) 0);
+            style = wb.getCellStyleAt(0);
         }
         StringBuilder sb = new StringBuilder();
         try (Formatter fmt = new Formatter(sb, Locale.ROOT)) {
